Added Mountain Theme
This theme uses this image taken from Microsoft: https://wallpaperhub.app/wallpapers/6064 that have been digitally modified to make it dynamic. Please note that these are not real images; only the one I have linked above is real, and the credit for that goes to Microsoft. I simply edited it to make it dynamic.