Added Borderlands 2 theme
Borderlands 2 Windshear Waste is the starter zone of the Borderlands 2 game, published by 2K and Gearbox. This set contains screenshots of the game taken at differents times of the day. Also, the game was in French when taken, so the display on the left bottom site will show French message from time to time.
This commit use jpg images instead of png to reduce the ddw file size, to overthrow Google Filesize limitations.